Change shellfuncs template and evaluservars
evaluservars now takes SHELL_PROVISIONER_ENV_FILE_NAME
and allows it be a user variable. This is so it is defined in one
place what .env file that the shell provisioners read from.

Also, the jq command in evaluservars now 'merges' the two json
objects using the * operator vs the + operator.